+++ package.json
@@ -104,14 +104,8 @@
   },
   "scripts": {
     "bootstrap": "lerna bootstrap",
-    "build": "yarn clean:build && lerna run prebuild && yarn build:babel && yarn build:ts && yarn build:tokens && lerna run postbuild",
-    "build:babel": "concurrently \"yarn build:babel:cjs\" \"yarn build:babel:esm\"",
-    "build:babel:cjs": "cross-env BABEL_ENV=production:cjs lerna exec --stream --ignore \"$(node ./scripts/getGlobPackages.js babel)\" --  babel src --out-dir dist/js",
-    "build:babel:esm": "cross-env BABEL_ENV=production:esm lerna exec --stream --ignore \"$(node ./scripts/getGlobPackages.js babel)\" -- babel src --out-dir dist/esm",
+    "build": "yarn clean:build && lerna run prebuild --parallel && lerna run build && lerna run postbuild",
     "build:docs": "yarn build && lerna run build --scope=@patternfly/react-docs --stream",
-    "build:tokens": "lerna run build --scope=@patternfly/react-tokens --stream",
-    "build:icons": "lerna run prebuild --scope=@patternfly/react-icons --stream",
-    "build:ts": "node ./scripts/copyTS.js",
     "clean": "yarn clean:build",
     "clean:build": "rm -rf packages/*/dist && rm -rf packages/patternfly-?/*/dist",
     "commit": "git-cz",
