+++ packages/patternfly-4/react-core/package.json
@@ -35,6 +35,11 @@
     "react-dom": "^15.6.2 || ^16.4.0"
   },
   "scripts": {
+    "build": "yarn build:babel; yarn build:ts",
+    "build:babel": "concurrently \"yarn build:babel:cjs\" \"yarn build:babel:esm\"",
+    "build:babel:cjs": "cross-env BABEL_ENV=production:cjs babel src --out-dir dist/js",
+    "build:babel:esm": "cross-env BABEL_ENV=production:esm babel src --out-dir dist/esm",
+    "build:ts": "node ./scripts/copyTS.js",
     "postbuild": "node ./build/copyStyles.js"
   },
   "optionalDependencies": {
@@ -43,6 +48,8 @@
   "devDependencies": {
     "@patternfly/patternfly-next": "1.0.43",
     "css": "^2.2.3",
-    "fs-extra": "^6.0.1"
+    "fs-extra": "^6.0.1",
+    "glob": "^7.1.2",
+    "npmlog": "^4.1.2"
   }
 }
